+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.166.2](https://github.com/adobe/spacecat-api-service/compare/v1.166.1...v1.166.2) (2025-08-28)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-utils to v1.50.0 ([#1200](https://github.com/adobe/spacecat-api-service/issues/1200)) ([8d2b010](https://github.com/adobe/spacecat-api-service/commit/8d2b0105ad053d241397b7481ba5763334bdbaf2))
+
+## [1.166.1](https://github.com/adobe/spacecat-api-service/compare/v1.166.0...v1.166.1) (2025-08-28)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.58.0 ([#1199](https://github.com/adobe/spacecat-api-service/issues/1199)) ([02bb36b](https://github.com/adobe/spacecat-api-service/commit/02bb36b5e5b1d77dff9ac10235f75e3f959dd170))
+
 # [1.166.0](https://github.com/adobe/spacecat-api-service/compare/v1.165.0...v1.166.0) (2025-08-27)